rename auth controller folder to authUser
